Окончание сортировки по файлам.
